Version Bump
Bumped the version and made a couple small text changes to this readme. The last contribution was really nice and there should be no worries about using the updated script in place of the last one. Nothing was broken or removed, only new features added/improved. See Contributions for details and credit!