Fixes an issue that prevented test from passing in IE.
The new menu's display is inline-block so that inline menus
won't take up all the space of their parent.

This caused tests that check that the menus topmost div's
display is block to fails in Internet Explorer, because
jQuery's .fadeIn() method didn't apply the display:block; css
to the menu's div.

This commit fixes the tests by checking visibility using the :visible
psudo instead.