Gives new class names to <td>s in the month-picker-year-table
The renaming is done to avoid conflicts with the users existing
CSS.
For example, the user might have defined a .next-year CSS
selector in there CSS which will mess up the menu.

It is also possible that users have applied their own CSS
to the menu to workaround some of the issues the CSS had
in version 2.x, in these cases I don't want the user's CSS to
mess up the new menu which works differently.

previous-year is now month-picker-previous
next-year is now month-picker-next
jump-years is now month-picker-title