Changes month <button> tag to <a> tag
Fixes a bug that caused the selected month to lose the
ui-state-active class when clicking outside the menu or
pressing any key (only in Chrome in with jQuery UI 1.11.4).

This change also fixes another issue (that also only happened
in Chrome) that caused the month button the user clicked on
to have an unnecessary blue "focus" ring around it.

Closes #50